[#1259] Remap download components and remove unused feature
