+++ frontend/src/components/index.js
@@ -31,4 +31,3 @@ export { default as InputAttributes } from "./InputAttributes";
 export { default as EntityFilters } from "./filters/EntityFilters";
 export { default as EntityDataFilters } from "./filters/EntityDataFilters";
 export { default as AttributeFilters } from "./filters/AttributeFilters";
-export { default as DownloadTable } from "./DownloadTable";
