+++ frontend/src/pages/downloads/components/index.js
@@ -1,0 +1,1 @@
+export { default as DownloadTable } from "./DownloadTable.jsx";
