+++ frontend/src/pages/downloads/style.scss
@@ -7,6 +7,9 @@
     .ant-card {
       margin-bottom: 20px;
     }
+    .download-filter {
+      font-size: 12px;
+    }
     .ant-layout {
       min-height: calc(100vh - 86px);
       &.site-layout {
