--- frontend/src/pages/export-data/style.scss
@@ -1,51 +1,0 @@
-@import "../../variables";
-
-#root {
-  #exportData {
-    .ant-table {
-      color: #707070;
-      padding: 20px;
-      .ant-table-cell {
-        padding: 24px 16px;
-      }
-      .ant-table-cell:first-child {
-        color: #00000099;
-        font-size: 28px;
-        img {
-          width: 32px;
-          height: 32px;
-        }
-      }
-      .ant-table-cell:nth-child(2) > div > div:first-child {
-        color: #707070;
-        margin-bottom: 3px;
-      }
-      .ant-table-cell:nth-child(2) > div > div:nth-child(3) {
-        margin-top: 6px;
-        .ant-tag {
-          border-style: dashed;
-        }
-      }
-
-      .ant-table-cell:last-child {
-        .ant-btn {
-          color: $blue;
-          text-transform: none;
-          font-weight: $font-weight-medium;
-          border-color: transparent;
-          background-color: transparent;
-          box-shadow: none;
-        }
-        .ant-btn:hover {
-          color: $blue;
-          border-color: transparent;
-        }
-      }
-      .ant-table-row.template {
-        .ant-table-cell:nth-child(2) > div > div:first-child {
-          color: $blue;
-        }
-      }
-    }
-  }
-}
